employee-app/requirements.txt
slack-bolt==1.18.0
python-dotenv==1.0.0
flask==2.3.3
requests==2.31.0
python-dateutil==2.8.2
pymongo==4.5.0
aiohttp==3.8.5
websockets==11.0.3 